update lbl module imports for PopulationFactory rename
