+++ radis/lbl/__init__.py
@@ -2,7 +2,7 @@
 """Core of the line-by-line calculations"""
 
 # prevent cyclic imports:
-from . import bands, base, broadening, calc, factory, labels, loader, overp
+from . import bands, broadening, calc, factory, labels, loader, overp, populations
 from .calc import calc_spectrum, spectrum_test
 from .factory import SpectrumFactory
 from .overp import LevelsList
